refactor: move feature specific imports into match arms for better readability
