+++ opentelemetry-jaeger/src/exporter/config/collector/http_client.rs
@@ -1,35 +1,5 @@
-#[cfg(feature = "surf_collector_client")]
-use async_trait::async_trait;
-#[cfg(any(
-    feature = "hyper_collector_client",
-    feature = "reqwest_blocking_collector_client",
-    feature = "reqwest_collector_client"
-))]
-use headers::authorization::Credentials;
-#[cfg(feature = "isahc_collector_client")]
-use isahc::config::Configurable;
 use opentelemetry_http::HttpClient as OtelHttpClient;
-#[cfg(feature = "surf_collector_client")]
-use std::convert::TryInto;
 use std::time::Duration;
-
-#[derive(Debug)]
-#[cfg(feature = "surf_collector_client")]
-struct BasicAuthMiddleware(surf::http::auth::BasicAuth);
-
-#[async_trait]
-#[cfg(feature = "surf_collector_client")]
-impl surf::middleware::Middleware for BasicAuthMiddleware {
-    async fn handle(
-        &self,
-        mut req: surf::Request,
-        client: surf::Client,
-        next: surf::middleware::Next<'_>,
-    ) -> surf::Result<surf::Response> {
-        req.insert_header(self.0.name(), self.0.value());
-        next.run(req, client).await
-    }
-}
 
 #[derive(Debug)]
 pub(crate) enum CollectorHttpClient {
@@ -49,7 +19,7 @@ pub(crate) enum CollectorHttpClient {
 
 impl CollectorHttpClient {
     // try to build a build in http client if users chose one. If none available return NoHttpClient error
-    #[allow(unused_variables)] // if the user enabled no build in client features. all parameters are unsed.
+    #[allow(unused_variables)] // if the user enabled no build in client features. all parameters are unused.
     pub(crate) fn build_client(
         self,
         collector_username: Option<String>,
@@ -62,7 +32,7 @@ impl CollectorHttpClient {
                 pipeline_name: "http_client",
                 config_name: "collector",
                 reason:
-                    "No http client provided. Consider enable one of the `surf_collector_client`, \
+                    "No http client provided. Consider enable one of the `hyper_collector_client`, `surf_collector_client`, \
         `reqwest_collector_client`, `reqwest_blocking_collector_client`, `isahc_collector_client` \
         features to use a build in http client. Or use `with_http_client` method in pipeline to \
         provide your own implementation."
@@ -70,6 +40,8 @@ impl CollectorHttpClient {
             }),
             #[cfg(feature = "isahc_collector_client")]
             CollectorHttpClient::Isahc => {
+                use isahc::config::Configurable;
+
                 let mut builder = isahc::HttpClient::builder().timeout(collector_timeout);
 
                 if let (Some(username), Some(password)) = (collector_username, collector_password) {
@@ -87,6 +59,9 @@ impl CollectorHttpClient {
             }
             #[cfg(feature = "surf_collector_client")]
             CollectorHttpClient::Surf => {
+                use std::convert::TryInto;
+                use opentelemetry_http::surf::BasicAuthMiddleware;
+
                 let client: surf::Client = surf::Config::new()
                     .set_timeout(Some(collector_timeout))
                     .try_into()
@@ -109,6 +84,8 @@ impl CollectorHttpClient {
             }
             #[cfg(feature = "reqwest_blocking_collector_client")]
             CollectorHttpClient::ReqwestBlocking => {
+                use headers::authorization::Credentials;
+
                 let mut builder =
                     reqwest::blocking::ClientBuilder::new().timeout(collector_timeout);
                 if let (Some(username), Some(password)) = (collector_username, collector_password) {
@@ -129,6 +106,8 @@ impl CollectorHttpClient {
             }
             #[cfg(feature = "reqwest_collector_client")]
             CollectorHttpClient::Reqwest => {
+                use headers::authorization::Credentials;
+
                 let mut builder = reqwest::ClientBuilder::new().timeout(collector_timeout);
                 if let (Some(username), Some(password)) = (collector_username, collector_password) {
                     let mut map = http::HeaderMap::with_capacity(1);
@@ -148,6 +127,7 @@ impl CollectorHttpClient {
             }
             #[cfg(feature = "hyper_collector_client")]
             CollectorHttpClient::Hyper => {
+                use headers::authorization::Credentials;
                 use opentelemetry_http::hyper::HyperClient;
 
                 let client = if let (Some(username), Some(password)) =
